Fix glitch when clicking tx
